+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.5] - 2026-09-17
+
+### Added
+
+- Add `FileOptions.skip_sync_file_validations` to skip synchronized file
+  validation during async checkpointing.
+
 ## [0.12.4] - 2026-08-12
 
 ### Fixed